Fix reactions serializer helper loading
